requirements.txt
streamlit
mysql-connector-python
pandas
numpy
plotly
seaborn
matplotlib
GitPython
python-dotenv
watchdog